Struct baum::Cursor

pub struct Cursor<'a, T> { /* private fields */ }

The cursor is a helper struct for operating on the tree without having to handle a lot of indices in the application itself. Only the entry node to the cursor has to be kept if it is not the root node.

Example

use baum::{Tree,Cursor};

let mut tree: Tree<i32> = Tree::new(1);
let mut cursor = Cursor::new_at_root(&mut tree);
cursor.add_child_and_go_down(2);
cursor.add_child(4);
cursor.add_child(5);
cursor.go_up();
cursor.add_child_and_go_down(3);
cursor.add_child(6);
cursor.add_child(7);

Implementations

impl<'a, T> Cursor<'a, T>

pub fn new(tree: &'a mut Tree<T>, idx: usize) -> Result<Self, TreeError>

Creates a new cursor for the given tree at the position of the passed index.

pub fn new_at_root(tree: &'a mut Tree<T>) -> Self

Creates a new cursor for the given tree at the root of the tree.

pub fn index(&self) -> usize

Retrieves the current index of the cursor position.

pub fn is_on_root(&self) -> bool

Checks whether or not the cursor is on the root node.

pub fn go_up(&mut self)

Moves the cursor up to the parent of the current node.

Example

The cursor focus is marked as (_).

Given the state:

   1
  / \
(2)  3

Then running

cursor.go_up();

Will result in this state:

  (1)
  / \
 2   3

pub fn add_child_and_go_down(&mut self, value: T) -> usize

Adds a new child to the current node and moves the cursor down to that node.

Example

The cursor focus is marked as (_).

Given the state:

   1
  / \
(2)  3

Then running

cursor.add_child_and_go_down(4);

Will result in this state:

     1
    / \
   2   3
  /
(4)

pub fn add_child(&mut self, value: T) -> usize

Adds a new child to the current node but keeps the cursor at the current node.

Example

The cursor focus is marked as (_).

Given the state:

   1
  / \
(2)  3

Then running

cursor.add_child(4);

Will result in this state:

     1
    / \
  (2)  3
  /
 4

pub fn value(&self) -> &T

Retrieves the value of the cursor at the current point.

Example

use baum::{Cursor,tr};
let mut tree = tr(1) / tr(2) / tr(3);
let mut cursor = Cursor::new_at_root(&mut tree);
let value = cursor.value();
assert_eq!(*value, 1);

This simple example showcases the retrival of the root value.

pub fn value_mut(&mut self) -> &mut T

Adds a new child to the current node but keeps the cursor at the current node.

Example

The cursor focus is marked as (_).

Given the state:

  (1)
  / \
 2   3

Then running a value_mut that increases the value by 2 will result in:

  (3)
  / \
 2   3

pub fn change_value(&mut self, val: T)

Puts the given value in place of the current position of the cursor.

pub fn children(&'a self) -> Vec<(&'a T, usize)>

Returns the values of all childrens in a vector.

Example

use baum::{Cursor,tr};
let mut tree = tr(1) / tr(2) / tr(3);
let mut cursor = Cursor::new_at_root(&mut tree);
let vals = cursor.children();
assert_eq!(vals.len(), 2);
assert_eq!(*vals[0].0, 2);
assert_eq!(*vals[1].0, 3);

pub fn destroy(self) -> usize

Destroys the cursor by consuming it and returns the index of the current node of the cursor.