Struct baum::Tree

pub struct Tree<T> {
    pub nodes: Vec<Node<T>>,
    pub root: usize,
}

Tree structure

Fields
nodes: Vec<Node<T>>root: usize
Implementations

impl<T> Tree<T>

pub fn new(inital: T) -> Self

A tree constructor. A tree can never be fully empty (use Option for that instead). Thus initializing a tree also comes with passing the inital root value:

Example

let tree: Tree<i32> = Tree::new(42);

pub fn root_value(&self) -> &T

pub fn get_root(&self) -> usize

Retrieves the index of the root node of the tree.

pub fn add_child(&mut self, parent: usize, value: T) -> Result<usize, TreeError>

Adds a child with the passed value to the parent node with the given index.

Example

This code snippet:

use baum::{Tree,Cursor};

let mut tree = Tree::new(1);
tree.add_child(tree.get_root(), 2);
tree.add_child(tree.get_root(), 3);

creates a tree of the following structure:

  1
 / \
2   3

pub fn remove_entry(&mut self, idx: usize) -> Result<(), TreeError>

Removes a child of the given index. Since any node in a tree can only have one parent, it is not necessary to also pass the parent node index.

pub fn traverse(&self) -> Vec<usize>

Traverses the tree.

Example

Given the tree:

        0
      /   \
     1     2
    / \   / \
   3   4 5   6

Produces the traversal: 0 1 3 4 2 5 6

pub fn move_node( &mut self, idx: usize, move_idx: isize ) -> Result<(), TreeError>

Moves the node at given index to the spot defined by the given move index. Either updates the position or returns an error for bad indexes.

Move indexes

Move indexes can be determined by the layer setup like this:

The first spot has move index 0

Given item with ID x, the spot beneath this item has move index x

Given a group with ID x, the first spot in this group has move index -x

Example

Given the tree:

        0
      /   \
     1     2
    /
   3

Then move indexes come like this (with the numbers in [] being the move indexes):

         0
       /   \
   [0]1[1] 2[2]
     /
[-1]3[3]

If we insert the value 4 at move_idx 1 then the tree becomes

         0
       / | \
      1  4  2
     /
    3

pub fn get_children(&self, idx: usize) -> Result<Vec<usize>, TreeError>

Returns all indices of the children of the provided node index.

pub fn get_parent(&self, idx: usize) -> Result<usize, TreeError>

Returns the parent node index of the given node index.

pub fn get_value(&self, idx: usize) -> Result<&T, TreeError>

Returns the value of the given node index.

pub fn value_mut(&mut self, idx: usize) -> Result<&mut T, TreeError>

pub fn set_value(&mut self, idx: usize, val: T) -> Result<(), TreeError>

Sets the value of the given node index.

pub fn add_tree_as_child( &mut self, other: Tree<T>, idx: usize ) -> Result<(), TreeError>

Adds a whole other tree as a child to the node provided as an index.